fix verifications that come to planning task page

+++ service/src/main/java/com/softserve/edu/service/calibrator/CalibratorPlanningTaskService.java
@@ -24,7 +24,7 @@ public interface CalibratorPlanningTaskService {
 
     void addNewTaskForTeam (Date taskDate, String serialNumber, List<String> verificationsId, String userId);
 
-    int findVerificationsByCalibratorEmployeeAndTaskStatusCount(String userName);
+    long findVerificationsByCalibratorEmployeeAndTaskStatusCount(String userName);
 
     Page<Verification> findByTaskStatusAndCalibratorId(Long Id, int pageNumber,
                                                        int itemsPerPage, String sortCriteria, String sortOrder);
